requirements.txt
torch
torchvision
webdataset
huggingface_hub
ninja
imageio
imageio-ffmpeg
Pillow
tqdm
numpy
